+++ CHANGELOG.md
@@ -430,10 +430,16 @@
 
 4.0.95 05/08/2025
 
-    Added MAT library condition properly in lix_tdo_v3.py
+    - Added MAT library condition properly in lix_tdo_v3.py
         # debug: find old files ended poorly
         # happened in some in cases in firmware version v4.1.23
         if mm[i:i+5] == b'\x00\x00\x00\x00\x00':
-    Changed this CHANGELOG.md file date format.
+
+    - Added the following to file: ble.py
+        # we don't want extra delays when doing orange macs
+        if _o:
+            delete_annotation(ev)
+    
+    - Changed this CHANGELOG.md file date format.